Add a pass to unify the accumulator and the input of the jax.scan loops.
This pass looks at the loops with accumulator patterns and unifies the
accumulation buffer with the input. The accumulation pattern usually comes
from jax.scan function. This transformation is beneficial in the cases where
the scan loop appears inside a loop body which causes a copy of the
accumulation buffer in the outer body.

PiperOrigin-RevId: 617588813