Add `MhloDynamicReshape` to XLA client API.
This adds a builder method which generates a `custom_call` to generate the MHLO op `dynamic_reshape`. Because of naming conflict with HLO's existing `DynamicReshape` (it's semantically different from `MhloDynamicReshape`), this op is prefixed with "Mhlo". I've also updated the API name for `DynamicBroadcastInDim` to `MhloDynamicBroadcastInDim` to match the names. These are only intended for export to MHLO or StableHLO, and cannot be compiled.

PiperOrigin-RevId: 625081376